pyproject.toml
[project]
name = "ims-api"
version = "0.1.0"
description = "Insurance Management System backend built with Django and DRF"
authors = [{ name = "IMS Team", email = "dev@ims.local" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0,<6",
    "djangorestframework>=3.14,<4",
    "django-environ>=0.11,<0.12",
    "drf-spectacular>=0.27,<0.28",
    "django-filter>=24.2,<25",
    "djangorestframework-simplejwt>=5.3,<6",
    "psycopg[binary]>=3.1,<4",
    "django-cors-headers>=4.3,<5",
    "whitenoise>=6.6,<7",
    "django-extensions>=3.2,<4",
    "gunicorn>=21.2,<22",
]

[project.optional-dependencies]
dev = [
    "black>=24.4,<25",
    "ruff>=0.3,<0.4",
    "isort>=5.13,<6",
    "mypy>=1.9,<2",
    "pytest>=8.1,<9",
    "pytest-django>=4.8,<5",
    "pytest-cov>=5.0,<6",
    "model-bakery>=1.17,<2",
    "coverage>=7.5,<8",
    "django-stubs>=4.2,<5",
    "djangorestframework-stubs>=3.14,<4",
]

[tool.black]
line-length = 100
target-version = ["py311"]
include = "\\.pyi?$"

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["apps", "config"]
combine_as_imports = true

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "DJ", "UP", "W", "B"]
ignore = ["DJ008"]
exclude = ["migrations"]

[tool.ruff.lint.per-file-ignores]
"config/settings/*" = ["E501"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--cov=apps --cov=config --cov-report=term-missing"

[tool.coverage.run]
branch = true
source = ["apps", "config"]
omit = [
    "*migrations*",
    "*/tests/*",
    "manage.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true

[mypy.plugins.django-stubs]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = ["dev"]
